+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add Zhang 2026 linezolid ([doi:10.3389/fphar.2026.1817282](https://doi.org/10.3389/fphar.2026.1817282)) -- critically ill children in a paediatric intensive care unit.
+
 - Add Padavia 2024 paracetamol and metabolites ([doi:10.1007/s40262-024-01439-3](https://doi.org/10.1007/s40262-024-01439-3)) -- extreme preterm neonates of 23-26 weeks' gestational age.
 
 - Add Beguin 2024 carboplatin ([doi:10.1186/s12917-024-04404-1](https://doi.org/10.1186/s12917-024-04404-1)) -- client-owned dogs with solid tumours, plus a thrombocyte-toxicity Emax model.